requirements.txt
streamlit
pygame
numpy
Pillow